+++ visit/src/util/mod.rs
@@ -1,1 +1,2 @@
+pub mod map;
 pub mod move_map;
